Disable LimitedModeJwtBearerGrantMockMvcTests
GE fork has its own JWT bearer implementation. Disable upstream tests.
This was disabled previously at the test level for both regular mode and
limited mode. But the merge commit previous to this commit moved the
disabling from individual test level to the class level to reduce
conflicts when tests are tweaked upstream, and to avoid having to
disable new tests added upstream. So these need to be disabled at the
class level for the limited mode test that extends the regular mode
tests.