Move Degraded JWT test from uaa/jwt/degraded to uaa/degraded
This moves the JWT specific degraded test into a location that is also
used in the predix_essentials branch. This makes it possible to use the
same exclude filter for degraded tests in both branches, which will
reduce conflicts when merging the JWT branch into predix_extensions or
rc branches. Update uaa/build.gradle to have the same exclude filters as
that used by the predix_extensions branch.