Modify JwtBearerGrantAT to support a use case when assertion token audience is different from access token issuer.
Assertion token is provided to UAA for authentication using JWT Bearer
grant flow. Access token is a token issued by UAA in exchange for the
assertion token after successful assertion token validation.
When assertion token is validated, its audience claim is compared to
the request URL. When access token is issued, its issuer claim is set
to issuer URI from uaa.yml config. JwtBearerGrantAT assumed that
audience and issuer are always the same. However, currently, for the
UAA deployment in AWS EKS, token request URL is set to a temporary
route [1] while token issuer is set to UAA route in CloudFoundry [2].

When DNS is setup to map UAA routes in EKS to CF routes, discrepancy
between audience of assertion token and issuer of access token should
go away.

- Add TOKEN_ISSUER_URL environment variable to be used by
JwtBearerGrantAT. This variable is set in acceptance test script for
EKS [3]. When set, this variable is used to assert access token issuer.
When not set, acceptance zone URL is used instead.

[1] https://github.devtools.predix.io/predix-security/uaa-k8s-deploy/blob/master/uaa/values-dev.yaml#L153
[2] https://github.devtools.predix.io/predix-security/uaa-k8s-deploy/blob/master/uaa/values-dev.yaml#L155
[3] https://github.devtools.predix.io/predix-security/uaa-k8s-deploy/commit/559aa8c9ebb719da4ff5824d0428af33de59131e