Update SAML auth MockMVC test to check IDP read from database
There are currently multiple redundant queries to read IDP data from the
identity_provider table for a zone when doing IDP discovery for SAML
login. Next commit will optimize this, but to see how many requests are made
currently, add assertions to an existing SAML authentication request
test about the number of times queries are made for IDPs to the
database.

- Update existing sendAuthnRequestToIdp test to add assertions after
  request to /saml/discovery, and request to the redirect URL after.