Implement Delete and PUT APIs
Update the controller to add delete and put API. Put API returns method
not allowed error since we are not implementing it.
Create a method for delete in the service class. As implemented in the
identity zone, implement ApplicationEventPublisherAware to obtain set
method of the event publisher. Get the identity zone object by name and
publish as a event as it is available in the identity zone API.
Add security config for delete and put API same as GET API.